Create Decentralized KYC Verification Process for Banks.md
Decentralized KYC Verification Process for Banks
Description 
Central Bank and other government banks face issues in tracking money laundering activities that are used for 
terrorism and other crimes. It is a threat to national security and is also adversely affecting the economy.
Background of the problem statement:
KYC (Know Your Customer) is a service provided by financial institutions such as banks. There are both public and 
private sector banks managed by a central bank. These banks are banned by the central bank from adding any new 
customer and do any more customer KYCs as they see suspicious activities that need to be sorted out first. Despite 
this, the banks add new customers and do the KYC in the background.
An immutable solution is needed where the central bank maintains a list of all the banks and tracks which banks are 
allowed to add new customers and perform KYC. It can also track which customer KYC is completed or pending along 
with customer details.
Banks can also add the new customer if allowed and do the KYC of the customers.